constant time

+++ src/main/java/frc/robot/Constants.java
@@ -57,7 +57,7 @@ public class Constants {
     public static final double kDriveTick2Feet = (1.0/128) * 6 * Math.PI * (1/12); //might not be right or needed
 
     //Autonomous
-    public static final int DRIVE_TIME = 4; 
+    public static final double DRIVE_TIME = 4.0; 
     public static final double AUTONOMOUS_DRIVE_PERCENTAGE = -.25;
 
 }
